+++ packages/block-editor/src/components/block-list/root-container.js
@@ -1,16 +1,15 @@
 /**
  * WordPress dependencies
  */
-import { useRef, createContext, useState } from '@wordpress/element';
+import { useRef, createContext } from '@wordpress/element';
 import { useSelect, useDispatch } from '@wordpress/data';
-import { Popover } from '@wordpress/components';
 
 /**
  * Internal dependencies
  */
 import useMultiSelection from './use-multi-selection';
 import { getBlockClientId } from '../../utils/dom';
-import BlockInsertionPoint from './insertion-point';
+import useInsertionPoint from './insertion-point';
 
 /** @typedef {import('@wordpress/element').WPSyntheticEvent} WPSyntheticEvent */
 
@@ -55,8 +54,6 @@ export default function RootContainer( { children, className } ) {
 		selectedBlockClientId,
 		hasMultiSelection,
 		isMultiSelecting,
-		getBlockRootClientId,
-		isBlockSelected,
 	} = useSelect( selector, [] );
 	const { selectBlock } = useDispatch( 'core/block-editor' );
 	const onSelectionStart = useMultiSelection( ref );
@@ -80,77 +77,20 @@ export default function RootContainer( { children, className } ) {
 		}
 	}
 
-	const [ isInserterShown, setIsInserterShown ] = useState( false );
-	const [ isInserterForced, setIsInserterForced ] = useState( false );
-	const [ inserterElement, setInserterElement ] = useState( null );
-	const [ inserterClientId, setInserterClientId ] = useState( null );
-	const [ inserterRootClientId, setInserterRootClientId ] = useState( null );
-
-	function onMouseMove( event ) {
-		if ( event.target.className !== className ) {
-			if ( isInserterShown ) {
-				setIsInserterShown( false );
-			}
-			return;
-		}
-
-		const rect = event.target.getBoundingClientRect();
-		const offset = event.clientY - rect.top;
-		const element = Array.from( event.target.children ).find( ( blockEl ) => {
-			return blockEl.offsetTop > offset;
-		} );
-
-		if ( ! element ) {
-			return;
-		}
-
-		const clientId = element.id.slice( 'block-'.length );
-
-		if ( ! clientId || isBlockSelected( clientId ) ) {
-			return;
-		}
-
-		const elementRect = element.getBoundingClientRect();
-
-		if ( event.clientX > elementRect.right || event.clientX < elementRect.left ) {
-			if ( isInserterShown ) {
-				setIsInserterShown( false );
-			}
-			return;
-		}
-
-		setIsInserterShown( true );
-		setInserterElement( element );
-		setInserterClientId( clientId );
-		setInserterRootClientId( getBlockRootClientId( clientId ) );
-	}
+	const { onMouseMove, InsertionPoint } = useInsertionPoint( {
+		className,
+		isMultiSelecting,
+		selectedBlockClientId,
+	} );
 
 	return <>
-		{ ( isInserterShown || isInserterForced ) && ! isMultiSelecting &&
-			<Popover
-				noArrow
-				animate={ false }
-				anchorRef={ inserterElement }
-				position="top right left"
-				focusOnMount={ false }
-				className="block-editor-block-list__block-popover"
-				__unstableSlotName="block-toolbar"
-			>
-				<BlockInsertionPoint
-					rootClientId={ inserterRootClientId }
-					clientId={ inserterClientId }
-					onFocus={ () => setIsInserterForced( true ) }
-					onBlur={ () => setIsInserterForced( false ) }
-					width={ inserterElement.offsetWidth }
-				/>
-			</Popover>
-		}
+		{ InsertionPoint && <InsertionPoint /> }
 		<div
 			ref={ ref }
 			className={ className }
 			onFocus={ onFocus }
 			onDragStart={ onDragStart }
-			onMouseMove={ ( isInserterForced || isMultiSelecting ) ? undefined : onMouseMove }
+			onMouseMove={ onMouseMove }
 		>
 			<Context.Provider value={ onSelectionStart }>
 				{ children }
